Updates for forcing refresh token and error counting
