+++ frontend/package.json
@@ -23,8 +23,8 @@
     "@material-ui/pickers": "^3.2.10",
     "@rematch/core": "^2.0.0-next.6",
     "@rematch/immer": "^1.2.0",
-    "@remote.it/components": "^1.1.12",
-    "@remote.it/services": "^0.9.8",
+    "@remote.it/components": "^1.1.17",
+    "@remote.it/services": "^0.9.10",
     "@remote.it/types": "^0.5.2",
     "@visx/axis": "^1.4.0",
     "@visx/event": "^1.3.0",
